Fixes #50: use knife vault for all commands
Based on the request for issue #50, this replaces the knife encrypt
and knife decrypt commands with knife vault.

This follows the convention of knife where you operate on an object with
commands (knife node edit, knife user edit, etc)

This also makes more sense because you're not creating an "encrypt" but
actually a "vault".