Add deauthorize client/admin to knife vault
This functionality was extracted from vault remove
because the remove command was doing too many
things. It was also confusing as to what you were
removing since it was possible to do many things
at once.

You are not allowed to deauthorize the default 
admin from the knife.rb file (node_name) since you
could create a situation where you were unable to
modify or delete the vault item.