Avoid sparse key read for non sparse secrets
Before this patch, every call to ChefVault::Item.load used to call the
chef server for sparse key twice:
- once to load keys
- once to decrypt symetrical key

This has a huge cost on secret read for nodes whose latency to the chef
server is high (we have example with ~300ms ping time).

This patch make sure we don't try to read sparse key when the secret is
marked as non sparse.

Change-Id: Ib0c493845ccd8d0630c00010b5f48a6a264f81dc
Signed-off-by: Grégoire Seux <g.seux@criteo.com>