Send a better error message when an invalid SSL key is encountered when creating a vault item
When creating a vault item we get an OpenSSL::PKey::RSAError error when the user or client
keys were generated from a version of chef-server before chef-server 12. The error raised
has no useful information for solving the problem. Pre chef-server 12
public keys are certificates instead of public keys. This change identifies the
user or client with an invalid public key. The key needs to be recreated, any existing
vault items used by the server need to be refreshed after the key is generated.

Signed-off-by: markgibbons <mark.gibbons@nordstrom.com>